license-admin/requirements.txt
fastapi
uvicorn[standard]
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
